requirements.txt
python-decouple==3.3
Flask~=1.1.2
mysql-connector-python~=8.0.21
pymongo~=3.11.0